@@ -1,11 +1,13 @@
 const express = require("express");
 const bodyParser = require('body-parser'); 
 const app = express();
-const routesHandler = require('./server/api/Mina_kurser')
+const routesHandler = require('./server/api/sok_kurser')
+const mina_kurser = require('./server/api/Mina_kurser')
 
 app.use(bodyParser.urlencoded({extended:false})); 
 app.use(bodyParser.json()); 
 app.use('/', routesHandler); 
+app.use('/', mina_kurser); 
 
 
 const PORT = process.env.PORT || 8080;
